Fix for Corran Horn spamming end-of-combat abillity turn after turn (Issue #135)
The initial event-based version of Corran Horn was supposed to be a
proof of concept, but I never got back to make it 100% rules-complete.
Adding a weapon disabled token (well, faking a weapon disabled token
that lasts to the next turn by applying two at different points) makes
Coran work correctly, and furthermore only blocks attacking on
subsequent turns if the player or AI *chooses* to use Corran's ability.
If the attack is skipped, Corran can attack next turn as per usual.