Further fix for Issue #110 (Guri softlocked game getting her focus from ability)
Somehow Guri is getting references to units that don't have the isdocked
member defined - possibly traversing all members of squadron is also
returning nearby objects - so making this check explicit and two-level
achieves the desired speedup of getrange vs. docked ships but also
avoids exceptions as seen in this issue.